feat: f37 PR -> removed depricated components + download functionality to LineChartOperations
